requirements.txt
bs4==0.0.1
Flask==0.12.2
gunicorn==19.7.1
line-bot-sdk==1.5.0
requests==2.18.4
